+++ packages/sources/itick/src/endpoint/quotes.ts
@@ -1,7 +1,7 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { TransportRoutes } from '@chainlink/external-adapter-framework/transports'
 import { config } from '../config'
-import { createAdapterResponseFromMessage } from '../transport/quote-shared'
+import { createAdapterResponseFromMessage } from '../transport/quotes'
 import { createHttpTransport } from '../transport/shared-http'
 import { createWsTransport } from '../transport/shared-ws'
 import { inputParameters } from './shared'
@@ -9,22 +9,26 @@ import { inputParameters } from './shared'
 export type BaseEndpointTypes = {
   Parameters: typeof inputParameters.definition
   Response: {
-    Result: number
+    Result: number | null
     Data: {
       symbol: string
-      lastPrice: number
+      askPrice: number
+      bidPrice: number
+      midPrice: number
+      askVolume: number
+      bidVolume: number
     }
   }
   Settings: typeof config.settings
 }
 
-const QUOTE_ENDPOINT_CONFIGS: { apiPath: string; name: string }[] = [
-  { apiPath: 'stock', name: 'stock-quote' },
-  { apiPath: 'indices', name: 'indices-quote' },
+const DEPTH_ENDPOINT_CONFIGS: { apiPath: string; name: string }[] = [
+  { apiPath: 'stock', name: 'stock-depth' },
+  { apiPath: 'indices', name: 'indices-depth' },
 ]
 
-export const endpoints = QUOTE_ENDPOINT_CONFIGS.map(({ apiPath, name }) => {
-  const type = 'quote'
+export const endpoints = DEPTH_ENDPOINT_CONFIGS.map(({ apiPath, name }) => {
+  const type = 'depth'
   const messageHandler = createAdapterResponseFromMessage
 
   return new AdapterEndpoint({
