+++ packages/sources/itick/src/transport/quotes.ts
@@ -1,6 +1,6 @@
 import { ProviderResult } from '@chainlink/external-adapter-framework/util'
 import { AdapterError } from '@chainlink/external-adapter-framework/validation/error'
-import { BaseEndpointTypes } from '../endpoint/depth'
+import { BaseEndpointTypes } from '../endpoint/quotes'
 
 export interface ResponseSchema {
   code: number // response code
