+++ packages/sources/itick/test/unit/quotes.test.ts
@@ -1,4 +1,4 @@
-import { createAdapterResponseFromMessage } from '../../src/transport/depth-shared'
+import { createAdapterResponseFromMessage } from '../../src/transport/quotes'
 
 describe('depth-shared', () => {
   const symbol = 'TEST'
